Optimize app.js
