+++ src/html/_header.htm
@@ -2,7 +2,7 @@
     <div class="header__row header__row_top top-header">
         <div class="top-header__container container">
             <nav class="top-header__menu menu-top-header">
-                <ul data-da=".menu__body, 991.98" class="menu-top-header__list">
+                <ul data-relocate=".menu__body, 991.98" class="menu-top-header__list">
                     <li class="menu-top-header__item menu-top-header__item_account"></li>
                     <li class="menu-top-header__item menu-top-header__item_catalog"><a data-catalog href="#" class="menu-top-header__link menu-top-header__link_catalog _icon-arrow_sh_d _ipl-after">Product catalog</a></li>
                     <li class="menu-top-header__item"><a href="#" class="menu-top-header__link">About us</a></li>
@@ -11,7 +11,7 @@
                     <li class="menu-top-header__item"><a href="#" class="menu-top-header__link">Contacts</a></li>
                 </ul>
             </nav>
-            <a class="top-header__user link-user _icon-user" data-da=".menu-top-header__item_account,991.98"><span>Account</span></a>
+            <a class="top-header__user link-user _icon-user" data-relocate=".menu-top-header__item_account,991.98"><span>Account</span></a>
             <nav class="menu">
                 <button class="menu__icon icon-menu">
                     <span></span>
@@ -26,7 +26,7 @@
     <div class="header__row header__row_body body-header">
         <div class="body-header__container container">
             <a href="./index.html" class="body-header__logo _icon-logo"></a>
-            <div data-da=".catalog-header__container, 479.98" class="body-header__search search-header">
+            <div data-relocate=".catalog-header__container, 479.98" class="body-header__search search-header">
                 <form action="#" class="search-header__form">
                     <button type="submit" tabindex="-1" class="search-header__button _icon-search"></button>
                     <input type="text" placeholder="Search.." name="form[]" autocomplete="off" data-error="Error" class="search-header__search">
@@ -48,8 +48,8 @@
                         <a class="number__order" href="tel:11234567890">Order a call</a>
                     </div>
                 </div>
-                <a data-da=".top-header__container, 991.98" href="#" class="information__fav _icon-fav"></a>
-                <a data-da=".top-header__container, 991.98" href="" class="information__shop shop">
+                <a data-relocate=".top-header__container, 991.98" href="#" class="information__fav _icon-fav"></a>
+                <a data-relocate=".top-header__container, 991.98" href="" class="information__shop shop">
                     <div href="#" class="shop__cart _icon-cart"></div>
                     <div class="shop__details">
                         <div class="shop__sum">0 $</div>
