+++ src/js/modules/script.js
@@ -1,11 +1,10 @@
 import { _slideDown, _slideUp } from "./functions.js";
-import { validateInit } from './validate.js';
-import { setViewPassword } from './forms.js'
 
 
-document.addEventListener("click", documentActions);
+document.addEventListener("click", documentClickActions);
+document.addEventListener('keydown', documentKeyActions);
 
-function documentActions(e){
+function documentClickActions(e){
     const targetElement = e.target;
     if (targetElement.closest('[data-parent]')){
         e.preventDefault();
@@ -49,18 +48,13 @@ function documentActions(e){
     }    
 }
 
-function subMenuBlockGridLayout() {
-    const menuBlocks = document.querySelectorAll(".sub-menu-catalog__block");
-    if (menuBlocks.length > 0){
-        menuBlocks.forEach(block => {
-            const categoryQty = block.querySelectorAll(".sub-menu-catalog__category").length;
-            block.classList.add(`sub-menu-catalog__block_${categoryQty}`);
-        })
+function documentKeyActions(e) {
+    if (e.key === "Enter" && e.target.matches('[tabindex="0"]')) {
+        e.target.click();
     }
 }
-subMenuBlockGridLayout();
 
-const burgerMenu = () =>{
+const burgerMenu = () => {
     const mql = matchMedia(`(max-width: ${991 / 16}em)`)
     const button = document.querySelector('.icon-menu');
     if (button) {
@@ -105,8 +99,7 @@ const burgerMenu = () =>{
 }
 burgerMenu();
 
-
-const handleHeader = () => {
+const handleStickHeader = () => {
     const topHeader = document.querySelector('.top-header');
     const bodyHeader = document.querySelector('.body-header');
     const catalogHeader = document.querySelector('.catalog-header');
@@ -178,8 +171,18 @@ const handleHeader = () => {
 
     
 }
-handleHeader();
+handleStickHeader();
 
+function subMenuBlockGridLayout() {
+    const menuBlocks = document.querySelectorAll(".sub-menu-catalog__block");
+    if (menuBlocks.length > 0){
+        menuBlocks.forEach(block => {
+            const categoryQty = block.querySelectorAll(".sub-menu-catalog__category").length;
+            block.classList.add(`sub-menu-catalog__block_${categoryQty}`);
+        })
+    }
+}
+subMenuBlockGridLayout();
 
 const handleSubMenuMore = () => {
     const submenu = document.querySelectorAll('.sub-menu-catalog__block');
@@ -203,166 +206,3 @@ const handleSubMenuMore = () => {
         })
     }
 }
-
-
-const handleCartHeight = () => {
-    const mq = window.matchMedia(`(min-width: ${992 / 16}em)`);
-    if (mq.matches){
-        const cartContainer = document.querySelector('.cart__container');
-        if (cartContainer) {
-            const headerHeight = document.querySelector(".header")?.scrollHeight;
-            const footerHeight = document.querySelector('.footer-cart')?.scrollHeight;
-            const windowHeight = window.innerHeight;
-            const cartContainerHeight = windowHeight - (headerHeight + footerHeight);
-            cartContainer.style.height = cartContainerHeight + 'px';
-        }
-    }
-}
-handleCartHeight();
-
-const setPlaceholders = () => {
-    const fields = document.querySelectorAll('.form__field:has(input:not([type="checkbox"]))');
-    fields.forEach(field => {
-        const input = field.querySelector('input');
-        if (input.value.length > 0) {
-            if (!field.classList.contains('_minimize-placeholder')) {
-                field.classList.add('_minimize-placeholder');
-            }
-        }
-        field.setAttribute('data-placeholder', input.placeholder ?? '');
-        input.addEventListener('focusin', function() {
-            if (!field.classList.contains('_minimize-placeholder')) {
-                field.classList.add('_minimize-placeholder');
-            }
-        });
-        input.addEventListener('focusout', function() {
-            if (input.value.length > 0) return;
-            if (field.classList.contains('_minimize-placeholder')) {
-                field.classList.remove('_minimize-placeholder');
-            }
-        });
-    });
-}
-setPlaceholders();
-
-const handleAuth = () => {
-    validateInit();
-    setViewPassword();
-    
-    const switchButton = document.querySelector('.auth__change-form');
-    if (!switchButton) {
-        return;
-    }
-    const params = new URLSearchParams(window.location.search);
-    let authForm = params.get('auth');
-    const redirectToParamString = params.get('r_link') ? '&r_link=' + params.get('r_link') : '';
-
-    if (authForm === null) {
-        window.history.pushState({}, "", window.location.origin + window.location.pathname + '?auth=login' + redirectToParamString);
-        authForm = 'login';
-    }
-    
-    const authBody = document.querySelector('.auth__body');
-    const formEl = authBody.querySelector('.auth__form');
-    const confirmPasswordParent = formEl.querySelector('.form__password-confirm').parentElement;
-    const dynamicChangeable = document.querySelectorAll('[data-changeable]');
-    authBody.classList.add(`_${authForm}`);
-
-    const revalidateEvent = new CustomEvent('revalidate');
-
-    const authFormText = {
-        login: {
-            switchFormButton: 'Register',
-            switchFormText: 'Don\'t have an account?',
-            submitFormButton: 'Log in',
-        },
-        register: {
-            switchFormButton: 'Log in',
-            switchFormText: 'Already have an account?',
-            submitFormButton: 'Register',
-        }
-    
-    }
-    
-    function renderFormInit() {
-        if (authForm === 'register') {
-            confirmPasswordParent.style.removeProperty('display');
-            dynamicChangeable.forEach(el => {
-                el.textContent = authFormText[authForm][el.dataset.name];
-            })
-        } else if (authForm === 'login') {
-            dynamicChangeable.forEach(el => {
-                el.textContent = authFormText[authForm][el.dataset.name];
-            })
-
-        }
-
-    }
-    renderFormInit();
-
-    switchButton.addEventListener('click', function() {
-        if (authForm === 'login'){
-            authBody.classList.remove('_login');
-            authBody.classList.add('_register');
-            authForm = 'register';
-            authFormAnimate(authForm);
-            window.history.pushState({}, '', window.location.origin + window.location.pathname  + "?auth=" + authForm + redirectToParamString);
-        } 
-        else if (authForm === 'register') {
-            authBody.classList.remove('_register');
-            authBody.classList.add('_login');
-            authForm = 'login';
-            authFormAnimate(authForm);
-            window.history.pushState({}, '', window.location.origin + window.location.pathname + "?auth=" + authForm + redirectToParamString);
-        }
-    })
-
-
-    function authFormAnimate(form) {
-        if (form === 'register') {
-            _slideDown(confirmPasswordParent, 250);
-        }
-        if (form === 'login') {
-            _slideUp(confirmPasswordParent, 250);
-        }
-        dynamicChangeable.forEach(el => {
-            el.classList.add('_fade_out');
-        });
-        setTimeout(() => {
-            dynamicChangeable.forEach(el => {
-                el.classList.remove('_fade_out');
-                el.classList.add('_fade_in');
-                el.textContent = authFormText[form][el.dataset.name];
-            })
-            document.dispatchEvent(revalidateEvent);
-        }, 250);
-    
-        setTimeout(() => {
-            dynamicChangeable.forEach(el => {
-                el.classList.remove('_fade_in');
-            })
-        }, 500);
-    }
-}
-
-handleAuth();
-
-const bindTabIndexElements = () => {
-    document.addEventListener('keydown', e => {
-        if (e.key === "Enter" && e.target.matches('[tabindex="0"]')) {
-            e.target.click();
-        }
-    });
-}
-
-bindTabIndexElements();
-
-
-const displayBenefitsWhyUsButton = () => {
-    const isMobile = document.documentElement.classList.contains('_touch');
-    const whyUs = document.querySelector('.advantages__title-mobile-arrow')
-    if (isMobile && whyUs) {
-        whyUs.classList.add('_show');
-    }
-}
-displayBenefitsWhyUsButton();
