--- src/js/modules/tippy.js
@@ -1,3 +1,0 @@
-import tippy from "tippy.js";
-
-tippy('[data-tippy-content]');
