+++ src/product.html
@@ -138,7 +138,7 @@
                                 </div>
                             </div>
                             <div class="body-product__actions actions-product">
-                                <div class="actions-product__price product-price product-price_big-card"><span class="product-price__actual-price">960 $</span><span class="product-price__old-price">740 $</span></div>
+                                <div class="actions-product__price product-price product-price_big-card"><span class="product-price__current">960 $</span><span class="product-price__old">740 $</span></div>
                                 <div class="actions-product__row">
                                     <div class="actions-product__quantity quantity">
                                         <button class="quantity__button quantity__button_minus" type="button"></button>
